requirements.txt
openslide-python
opencv-python
scikit-image
matplotlib
tqdm